requirements.txt
google-adk
google-genai
pandas
ipython
notebook
pytest
